refactor(agent): extract communication module from config.rs
Move CommunicationConfig, check_communication_allowed, and
check_max_depth to a new agent/communication.rs module to reduce
config.rs from 536 to 418 lines (limit 500).

Source: issue #223
Type: refactor
